Share audio test base fixture. Disable song tests on all platforms but Desktop VK and Desktop DX12.
