+++ Tests/Framework/Audio/SoundEffectTest.cs
@@ -13,15 +13,8 @@ using Microsoft.Xna.Framework.Graphics;
 namespace MonoGame.Tests.Audio
 {
     [Category("Audio")]
-    public class SoundEffectTests
+    public class SoundEffectTests : AudioTestFixtureBase
     {
-        [SetUp]
-        public void Setup()
-        {
-            // Necessary to get audio initialised
-            FrameworkDispatcher.Update();
-        }
-
         [Test]
         public void Statics()
         {
@@ -454,36 +447,6 @@ namespace MonoGame.Tests.Audio
         }
 #endif
 
-        // Proxy for the content manager used in SoundEffectFromContent
-        class GraphicsDeviceProxy : IGraphicsDeviceService
-        {
-            public GraphicsDevice GraphicsDevice
-            {
-                get { return null; }
-            }
-
-            public event EventHandler<EventArgs> DeviceCreated;
-
-            public event EventHandler<EventArgs> DeviceDisposing;
-
-            public event EventHandler<EventArgs> DeviceReset;
-
-            public event EventHandler<EventArgs> DeviceResetting;
-        }
-
-        class ContentManagerProxy : ContentManager
-        {
-            public ContentManagerProxy(IServiceProvider services): base(services) {}
-
-            protected override Stream OpenStream(string assetName)
-            {
-                var fileName = Path.Combine(RootDirectory, assetName + ".xnb");
-                if (File.Exists(fileName))
-                    return new FileStream(fileName, FileMode.Open, FileAccess.Read);
-                return base.OpenStream(assetName);
-            }
-        }
-
         [TestCase("tone_mono_44khz_8bit", 5000000)]
         [TestCase("tone_stereo_44khz_8bit", 5000000)]
         [TestCase("tone_mono_44khz_16bit", 5000000)]
@@ -502,10 +465,7 @@ namespace MonoGame.Tests.Audio
 #endif
         public void SoundEffectFromContent(string filename, long durationTicks)
         {
-            var services = new GameServiceContainer();
-            services.AddService<IGraphicsDeviceService>(new GraphicsDeviceProxy());
-            var content = new ContentManagerProxy(services);
-            var soundEffect = content.Load<SoundEffect>(Paths.Audio(filename));
+            var soundEffect = _content.Load<SoundEffect>(Paths.Audio(filename));
             Assert.AreEqual(durationTicks, soundEffect.Duration.Ticks);
         }
     }
